Don't try to delete directory that may have disappeared

+++ BleemSync.Extensions.PlayStationClassic/BleemSync.Extensions.PlayStationClassic.Core/Services/GameManagerService.cs
@@ -191,7 +191,7 @@ namespace BleemSync.Extensions.PlayStationClassic.Core.Services
         {
             // Delete files first
             var gameDirectory = Path.Combine(_baseGamesDirectory, node.Id.ToString());
-            Directory.Delete(gameDirectory, true);
+            if (Directory.Exists(gameDirectory)) Directory.Delete(gameDirectory, true);
 
             var game = _context.Games.Find(node.Id);
